Finish the presentation layer| enhance on all layers

+++ src/main/kotlin/domain/services/PaymentService.kt
@@ -2,10 +2,11 @@ package domain.services
 
 import domain.models.Payment
 import domain.models.utils.PaymentMethod
+import java.math.BigDecimal
 
-
+// Interface for payment service operations
 interface PaymentService {
-    suspend fun makePayment(orderId: String, amountPaid: Double, method: PaymentMethod): Boolean
+    suspend fun makePayment(orderId: String, amountPaid: BigDecimal, method: PaymentMethod): Boolean
     suspend fun getPaymentById(id: String): Payment?
     suspend fun getPaymentsForOrder(orderId: String): List<Payment>
     suspend fun listAllPayments(): List<Payment>
